models documentation: add note about validation dict key names
Add a note to the model documentation that explicitly calls out the pattern used by the validation dict. The keys are *not* the field names; rather, they follow the pattern `{field name}_validator`. This can save users a bit of time if they don't read the code sample very carefully (like me!).